columns_upgrades

+++ src/main/java/com/vaadin/domain/ItemFinalize.java
@@ -11,6 +11,6 @@ import lombok.Setter;
 @Getter
 @Setter
 public class ItemFinalize {
-    Double idValue;
+    Long idValue;
     WalletDto walletDto;
 }
